+++ include/ignition/transport/TransportTypes.hh
@@ -42,6 +42,7 @@ namespace ignition
     class IRepHandler;
     class IReqHandler;
     class ISubscriptionHandler;
+    class MessageInfo;
 
     /// \def Addresses_M
     /// \brief Map that stores all generic publishers.
@@ -115,9 +116,20 @@ namespace ignition
     using SrvDiscoveryCallback =
       std::function<void(const ServicePublisher &_publisher)>;
 
+    /// \def MsgCallback
+    /// \brief User callback used for receiving messages:
+    ///   \param[in] _msg Protobuf message containing the topic update.
+    ///   \param[in] _info Message information (e.g.: topic name).
+    template <typename T>
+    using MsgCallback =
+      std::function<void(const T &_msg, const MessageInfo &_info)>;
+
     /// \def Timestamp
     /// \brief Used to evaluate the validity of a discovery entry.
     using Timestamp = std::chrono::steady_clock::time_point;
+
+    /// \brief The string type used for generic messages.
+    const std::string kGenericMessageType = "google.protobug.Message";
   }
 }
 #endif