+++ src/Node.cc
@@ -27,6 +27,7 @@
 #include <unordered_set>
 #include <vector>
 
+#include "ignition/transport/MessageInfo.hh"
 #include "ignition/transport/Node.hh"
 #include "ignition/transport/NodeOptions.hh"
 #include "ignition/transport/NodePrivate.hh"
@@ -165,6 +166,21 @@ bool Node::Publisher::Valid() const
 }
 
 //////////////////////////////////////////////////
+bool Node::Publisher::HasConnections() const
+{
+  ISubscriptionHandlerPtr firstSubscriberPtr;
+  auto &publisher = this->dataPtr->publisher;
+
+  std::lock_guard<std::recursive_mutex> lk(this->dataPtr->shared->mutex);
+
+  return this->Valid() &&
+    (this->dataPtr->shared->localSubscriptions.FirstHandler(
+       publisher.Topic(), publisher.MsgTypeName(), firstSubscriberPtr) ||
+     this->dataPtr->shared->remoteSubscribers.HasTopic(
+       publisher.Topic(), publisher.MsgTypeName()));
+}
+
+//////////////////////////////////////////////////
 bool Node::Publisher::Publish(const ProtoMsg &_msg)
 {
   if (!this->Valid())
@@ -193,13 +209,21 @@ bool Node::Publisher::Publish(const ProtoMsg &_msg)
 
     hasLocalSubscribers = this->dataPtr->shared->localSubscriptions.Handlers(
       this->dataPtr->publisher.Topic(), handlers);
-    hasRemoteSubscribers =this->dataPtr->shared->remoteSubscribers.HasTopic(
-      this->dataPtr->publisher.Topic());
+    hasRemoteSubscribers = this->dataPtr->shared->remoteSubscribers.HasTopic(
+      this->dataPtr->publisher.Topic(), _msg.GetTypeName());
   }
 
   // Local subscribers.
   if (hasLocalSubscribers)
   {
+    // Get the topic and remove the partition name.
+    std::string topic = this->dataPtr->publisher.Topic();
+    topic.erase(0, topic.find_last_of("@") + 1);
+
+    // Create and populate the message information object.
+    MessageInfo info;
+    info.SetTopic(topic);
+
     for (auto &node : handlers)
     {
       for (auto &handler : node.second)
@@ -208,10 +232,13 @@ bool Node::Publisher::Publish(const ProtoMsg &_msg)
 
         if (subscriptionHandlerPtr)
         {
-          if (subscriptionHandlerPtr->TypeName() != _msg.GetTypeName())
+          if (subscriptionHandlerPtr->TypeName() != kGenericMessageType &&
+              subscriptionHandlerPtr->TypeName() != _msg.GetTypeName())
+          {
             continue;
+          }
 
-          subscriptionHandlerPtr->RunLocalCallback(_msg);
+          subscriptionHandlerPtr->RunLocalCallback(_msg, info);
         }
         else
         {
@@ -415,6 +442,11 @@ bool Node::Unsubscribe(const std::string &_topic)
       msg.rebuild(this->dataPtr->nUuid.size());
       memcpy(msg.data(), this->dataPtr->nUuid.data(),
              this->dataPtr->nUuid.size());
+      socket.send(msg, ZMQ_SNDMORE);
+
+      msg.rebuild(kGenericMessageType.size());
+      memcpy(msg.data(), kGenericMessageType.data(),
+             kGenericMessageType.size());
       socket.send(msg, ZMQ_SNDMORE);
 
       std::string data = std::to_string(EndConnection);